Remove Guide Book from Cyclic. Many items are missing / recipes changed
So, I would rather not have people ask me why there are missing items and if it's a bug. Because people will be like that. It happens all the time. So, it won't be like that anymore. Now people will just ask where the book is. Also, people will probably still ask where EnderIO and Sieving is!