fix: double announcement of LostCities advancement
Because there is an alternate description for Lost and Found for players who start in the Lost Cities world, a player travelling to the Lost Cities will always receive both advancements. This commit removes the announcement for the lost cities start variant so players are only notified of the advancement once.

It also removes the reward so players will only receive one overworld cake.

Closes Darkosto/SkyFactory4Issues#466